Show Coral Talk for articles published after migration start date
For articles published after a certain date (which is not decided yet) we will start using Coral Talk.

This relies on two environment variables named COMMENTS_USE_CORAL_TALK and COMMENTS_USE_CORAL_MILESTONE_DATE.

COMMENTS_USE_CORAL_TALK needs to be set to `true` for this to work.

COMMENTS_USE_CORAL_MILESTONE_DATE defines at which point we start using Coral Talk. Articles published before that date will be rendered with Livefyre, and those that are published after that date will be rendered with Coral.

If COMMENTS_USE_CORAL_TALK is not true or COMMENTS_USE_CORAL_MILESTONE_DATE is not set, the page will render with Livefyre.