Remove comment count from Blogs stream page
We're unable to render both the Coral and Livefyre comment counts
on the Blogs stream page because `o-comments` v5 conflicts with v6.

While we are in the process of migrating to Coral, we are temporarily
removing the comment count. Once migration is complete, we will render
the Coral comment count.