Load MarketsLive homepage content from file system
The content needed for the Markets Live homepage is already available in
the archived Markets Live sessions so there's no need to get that info
from alphaville-es-interface-service.

This work came out of the need to remove the Markets Live API from
alphaville-es-interface-service. Fingers crossed, it should now be
easier to do that.

This work is related to the ongoing decommissioning of Markets Live,
see https://github.com/Financial-Times/next/issues/400.