Remove the use of the email platform
The email platform are deprecating the support for their endpoint that
returns user data.

This service doesn't require this endpoint anyway as we can match
based on uuid rather than worrying about matching based on email
address.

This also has the benefits of not needing to call a service and also not
needing to contect to a database to receieve the allowed users.